// src/types/Order.tsx
export enum OrderStatus {
    PENDING = "PENDING",
    DELIVERED = "DELIVERED",
    CANCELED = "CANCELED",
    ONGOING = "ONGOING",
  }

  export enum ShippingType {
    FREE = "FREE",
    EXPRESS = "EXPRESS",
    FLAT = "FLAT",
  } 
  
  export enum PaymentMethod {
    CASH = "CASH", //means payment on delivery
  }